svg/pservers/scripted/stop-color-inheritance-currentcolor.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" color="red"><title>SVG Paint Servers: &apos;stop-color&apos; inheritance of &apos;currentcolor&apos;</title><metadata><h:link href="https://svgwg.org/svg2-draft/pservers.html#StopColorProperty" rel="help"/><h:link href="https://drafts.csswg.org/css-color/#resolve-color-values" rel="help"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><linearGradient id="g" stop-color="currentcolor"><stop color="green" stop-color="inherit"/></linearGradient><script><![CDATA[

    test(function() {
      let stop = document.querySelector('stop');
      assert_equals(getComputedStyle(stop).stopColor, 'rgb(0, 128, 0)');
    }, '"currentcolor" is inherited as a keyword');

  ]]></script></svg>
