LayoutTests/imported/w3c/web-platform-tests/svg/pservers/scripted/stop-color-inheritance-currentcolor-expected.txt


FAIL "currentcolor" is inherited as a keyword assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
